extract GroupPage ToggleButton section into its own component
